requirements.txt
mysql-connector-python==8.0.31
python-dotenv==1.0.0
Flask>=2.0
Werkzeug>=2.0
